Use a mapper to handle the CSRF token response
