+++ ui/src/maps/csrfToken.ts
@@ -1,0 +1,11 @@
+import { CsrfToken } from "@/models/CsrfToken"
+import { CsrfTokenResponse } from "@/types/csrfTokenResponse"
+
+export const mapCsrfTokenResponseToCsrfToken = (response: CsrfTokenResponse): CsrfToken => {
+    return {
+        token: response.token,
+        expiration: new Date(response.expiration),
+        issued: new Date(),
+    }
+}
+
