+++ ui/src/models/CsrfToken.ts
@@ -1,0 +1,5 @@
+export type CsrfToken = {
+    token: string
+    expiration: Date
+    issued: Date
+}
