+++ ui/src/types/csrfTokenResponse.ts
@@ -1,0 +1,5 @@
+export type CsrfTokenResponse = {
+    token: string
+    expiration: Date
+}
+
